Proyecto1/src/main/java/com/proyecto1/ecommerce/data/ProductoData.java
package com.proyecto1.ecommerce.data;

import java.util.List;

import javax.sql.DataSource;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.jdbc.core.BeanPropertyRowMapper;
import org.springframework.jdbc.core.JdbcTemplate;
import org.springframework.stereotype.Repository;
import org.springframework.transaction.annotation.Transactional;

import com.proyecto1.ecommerce.domain.Producto;
import com.proyecto1.ecommerce.domain.Rol;

@Repository
public class ProductoData {

	@Autowired
	private JdbcTemplate jdbcTemplate;
	@Autowired
	private DataSource dataSource;
	
	@Transactional(readOnly = true)
	public List<Producto> findAll() {
		String sqlSelect = "CALL `Producto_SelectAll`();";
		return jdbcTemplate.query(sqlSelect, new BeanPropertyRowMapper(Producto.class));
	}
	
}
